+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.0.3-alpha.8",
-  "contentSha256_12": "ebe1ea39bf75",
+  "pkgVersion": "1.0.3-alpha.9",
+  "contentSha256_12": "eb99a085de19",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.0.3-alpha.8";
-export const contentSha256_12 = "ebe1ea39bf75";
+export const pkgVersion = "1.0.3-alpha.9";
+export const contentSha256_12 = "eb99a085de19";
 export const file = "nodalia-cards.bundle.js";